Move the D-Bus event handler into the dbus_api module
For better encapsulation.

Also, make consts module private, since the constants no longer have to
be exported.
